Full implementation of LED sequence
LED sequence implemented with ~0.1s resolution.
Disabled interrupts in IRQ handler to save stack space and have
higher consistency, plus we don't have the LED on while doing
touch sense anyway.

When the user first touches the touch pad, the LED will flash to
indicate the start of the programming sequence.  After this, touches
are recorded for 10 seconds, after which the recorder sequence will
be played back over and over.